Moved apis/constants.js into helpers subfolder and renamed it to endpoints.js
