+++ app/lib/api/apis/confirmation.js
@@ -6,7 +6,7 @@
 import * as m from '../model';
 import ApiRequest from '../ApiRequest';
 
-import { CONFIRMATION_ENDPOINT } from './constants';
+import { CONFIRMATION_ENDPOINT } from './helpers/endpoints';
 
 const post = (
   email: string,
