+++ app/lib/api/apis/notifications.js
@@ -6,7 +6,7 @@
 import * as m from '../model';
 import ApiRequest from '../ApiRequest';
 
-import { NOTIFICATIONS_ENDPOINT } from './constants';
+import { NOTIFICATIONS_ENDPOINT } from './helpers/endpoints';
 
 const getAll = async (): Promise<m.ApiResponseData> => {
   const request = new ApiRequest();
