+++ app/lib/api/apis/topics.js
@@ -9,11 +9,7 @@ import contentItems from 'modules/contentItems';
 import * as m from '../model';
 import ApiRequest from '../ApiRequest';
 
-import {
-  USERS_ENDPOINT,
-  TOPICS_ENDPOINT,
-  TOPICS_CONTENT_ENDPOINT,
-} from './constants';
+import { USERS_ENDPOINT, TOPICS_ENDPOINT, TOPICS_CONTENT_ENDPOINT } from './helpers/endpoints';
 
 const destroy = (
   id: Identifier,
